Workaround issue with port forwarding the RHDH port
There is a limitation with networks that have "internal: true": ports are not published with Docker, but are with Podman.
So with Docker, port forwarding does not work at all from the host, because the host port is not exposed.
See https://github.com/moby/moby/issues/36174
As a workaround, we are adding it to the default network,
but note that there is nothing preventing requests from bypassing the proxy to reach the outside.