Add quickstart guides for users and admins
Added quickstart guides for users and admins in app-config.yaml. This PR will reconfigure the Quickstart docked menu so that it offers a broad selection of suggestions for new users (both admins and developers) to try in order to familiarize themselves with Red Hat Developer Hub and its major features. There is one optional (commented) item intended for users of Developer Lightspeed.